Rename file with Message class from jobs->message.py
Renames the module with Message class from jobs to message.py, this
follows PEP8 and prevents me constantly getting confused on where the
message implementation has gone
