+++ queue_processors/autoreduction_processor/autoreduction_processor.py
@@ -15,7 +15,7 @@ import stomp
 
 from twisted.internet import reactor
 
-from model.message.job import Message
+from model.message.message import Message
 from queue_processors.autoreduction_processor.autoreduction_logging_setup import logger
 # pylint:disable=no-name-in-module,import-error
 from queue_processors.autoreduction_processor.settings import MISC
