+++ queue_processors/queue_processor/queueproc_utils/messaging_utils.py
@@ -10,7 +10,7 @@ Utils module for sending messages to queues.
 
 # pylint: disable=cyclic-import
 from model.database import access
-from model.message.job import Message
+from model.message.message import Message
 from queue_processors.queue_processor.settings import FACILITY
 
 from utils.clients.queue_client import QueueClient
